LUCENE-1257: Add generics to highlighter and more for benchmark
git-svn-id: https://svn.apache.org/repos/asf/lucene/java/trunk@829524 13f79535-47bb-0310-9956-ffa450edef68

+++ contrib/benchmark/src/java/org/apache/lucene/benchmark/byTask/tasks/OptimizeTask.java
@@ -41,7 +41,7 @@ public class OptimizeTask extends PerfTask {
 
   public void setParams(String params) {
     super.setParams(params);
-    maxNumSegments = (int) Double.valueOf(params).intValue();
+    maxNumSegments = Double.valueOf(params).intValue();
   }
 
   public boolean supportsParams() {
